+++ packages/adapters/src/harmony/index.ts
@@ -2,10 +2,10 @@
  * TaroViz 鸿蒙OS适配器
  * 基于鸿蒙OS应用的Canvas组件实现图表渲染
  */
+import { Adapter, HarmonyAdapterOptions, EChartsOption, EventHandler } from '@agions/taroviz-core';
+import { uuid } from '@agions/taroviz-core';
 import { Canvas } from '@tarojs/components';
 import Taro from '@tarojs/taro';
-import { Adapter, HarmonyAdapterOptions, EChartsOption, EventHandler } from '@taroviz/core/types';
-import { uuid } from '@taroviz/core/utils';
 import * as echarts from 'echarts/core';
 import React from 'react';