+++ packages/adapters/src/lark/index.ts
@@ -4,8 +4,8 @@
  */
 import { Canvas } from '@tarojs/components';
 import Taro from '@tarojs/taro';
-import { Adapter, MiniAppAdapterOptions, EChartsOption, EventHandler } from '@taroviz/core/types';
-import { uuid } from '@taroviz/core/utils';
+import { Adapter, MiniAppAdapterOptions, EChartsOption, EventHandler } from '@agions/taroviz-core';
+import { uuid } from '@agions/taroviz-core';
 import * as echarts from 'echarts/core';
 import React from 'react';