+++ packages/adapters/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/Agions/TaroViz/compare/v1.0.2...v1.0.3) (2025-04-14)
+
+### Bug Fixes
+
+- 修复多个适配器中使用`Function`类型的问题，替换为更具体的`EventHandler`类型
+- 优化WeappAdapter类型定义，添加ExtendedWeappAdapterOptions接口支持component属性
+- 修复其他linter警告和代码规范问题
+
 ## [1.0.2](https://github.com/Agions/TaroViz/compare/v1.0.1...v1.0.2) (2025-04-12)
 
 **Note:** Version bump only for package @agions/taroviz-adapters
@@ -13,4 +21,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.5.1](https://github.com/Agions/TaroViz/compare/v0.4.0...v0.5.1) (2025-04-11)
 
-**Note:** Version bump only for package @taroviz/adapters
+**Note:** Version bump only for package @agions/adapters